+++ scripts/test-coverage/merge.js
@@ -1,7 +1,6 @@
 const path = require('path')
 const fsSync = require('fs')
 
-
 const CoverageDirectory = '/coverage'
 
 function getRelevantPackages() {
@@ -9,7 +8,7 @@ function getRelevantPackages() {
     .readdirSync(path.join(__dirname, '../../packages'))
     .filter(pkg => {
       return fsSync.existsSync(
-        path.join(__dirname, '../../packages', pkg,CoverageDirectory )
+        path.join(__dirname, '../../packages', pkg, CoverageDirectory)
       )
     })
 
@@ -17,23 +16,28 @@ function getRelevantPackages() {
 }
 
 function getCoverageDirectories(packages) {
-    const webPackages =  packages.map(pkg => {
-        return path.join(__dirname, '../../packages', pkg, CoverageDirectory , 'web')
-    })
-    const nativePackages =  packages.map(pkg => {
-        return path.join(__dirname, '../../packages', pkg, CoverageDirectory , 'web')
-    })
-    return [...webPackages, ...nativePackages]
+  const webPackages = packages.map(pkg => {
+    return path.join(__dirname, '../../packages', pkg, CoverageDirectory, 'web')
+  })
+  const nativePackages = packages.map(pkg => {
+    return path.join(
+      __dirname,
+      '../../packages',
+      pkg,
+      CoverageDirectory,
+      'native'
+    )
+  })
+  return [...webPackages, ...nativePackages]
 }
 
 async function main() {
-    const packages = getRelevantPackages()
-    const coverageDirectories = getCoverageDirectories(packages)
+  const packages = getRelevantPackages()
+  const coverageDirectories = getCoverageDirectories(packages)
 
+  const mergeCoverage = await import('@monorepo-template/merge-coverage')
 
-    const mergeCoverage = await import('@monorepo-template/merge-coverage')
-
-    await mergeCoverage.default(coverageDirectories);
+  await mergeCoverage.default(coverageDirectories)
 }
 
 main()