+++ packages/a11y/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @interunit/a11y
 
+## 0.0.20
+
+### Patch Changes
+
+- Updated dependencies
+  - @interunit/primitives@0.0.22
+
 ## 0.0.19
 
 ### Patch Changes