+++ packages/modal/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @interunit/modal
 
+## 0.0.18
+
+### Patch Changes
+
+- Improvements from testing and other small adjustments.
+- Updated dependencies
+  - @interunit/primitives@0.0.22
+  - @interunit/a11y@0.0.20
+
 ## 0.0.17
 
 ### Patch Changes