+++ packages/primitives/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @interunit/primitives
 
+## 0.0.22
+
+### Patch Changes
+
+- Improvements from testing and other small adjustments.
+
 ## 0.0.21
 
 ### Patch Changes